Windows support: build all 5 backends + Python bindings, fix runtime loading
Build system
- Convert the ROCm and OneAPI backends from CMake's native HIP/SYCL
  language support to a direct custom-command pattern (hipcc.exe/icpx.exe
  invoked explicitly), since CMake hard-blocks mixing MSVC with
  Clang-family languages on Windows.
- Fix dozens of MSVC portability issues surfaced by building the CPU
  backend, core library, and Python bindings for the first time on
  Windows: GCC/Clang-only __attribute__((target(...))) function
  multiversioning replaced with a portable macro, __restrict__ ->
  __restrict, unsigned OpenMP loop counters -> signed (required by
  MSVC's OpenMP 2.0 implementation), __builtin_mul_overflow/
  __builtin_add_overflow -> tenzor::detail::checked_*_overflow,
  POSIX-only headers/APIs guarded or replaced across pinned_allocator,
  rnn_kernels, and others.
- Python bindings now build and run: pybind11 wiring, MSVC-specific
  static-initializer trick for OpenMP thread configuration (.CRT$XCU
  section, replacing __attribute__((constructor))), ssize_t alias fix.

Runtime DLL loading (Windows)
- BackendLoader::load_library() used bare LoadLibraryA, which silently
  ignores AddDllDirectory-registered search paths -- switched to
  LoadLibraryExW with LOAD_LIBRARY_SEARCH_DEFAULT_DIRS.
- Added ensure_win_dll_search_setup() (src/backend/loader.cpp): registers
  this build's vendor SDK directories and pre-loads the specific runtime
  DLLs (Intel oneAPI's SYCL runtime, oneMKL SYCL interop) that fail to
  resolve their own nested dependencies when loaded cold even with the
  right search paths registered. Runs once, lazily, before any backend
  load -- fixes backend loading for both Python and plain C++ executables
  linking tenzor_core, with no caller-side setup required.
- CMakeLists.txt generates the vendor directory list once from this
  build's actual discovered SDK paths, shared between a C++ header
  (build/generated/tenzor_win_dll_dirs.generated.hpp, consumed by
  loader.cpp) and a Python manifest (build/python/tenzor/_win_dll_dirs.py,
  consumed by __init__.py's Windows bootstrap -- needed because Python
  must register directories before tenzor_core.dll itself has loaded).
- create_backend() and register_kernels() in the ROCm/OneAPI backends
  lacked __declspec(dllexport): CMAKE_WINDOWS_EXPORT_ALL_SYMBOLS only
  applies to normal add_library(SHARED) targets, not these raw
  custom-command builds. GetProcAddress silently failed for both.

Backend/runtime correctness fixes (found while getting Windows working,
apply on all platforms)
- BackendLoader::load_backend()'s failure paths called unload_library()
  right after a backend constructor partially ran -- unsafe (matches this
  file's own existing shutdown()/~BackendLoader() reasoning: a backend
  library's transitive dependencies may register their own exit/unload
  state). Confirmed via a minimal repro: FreeLibrary() right after a
  failed Vulkan vkCreateInstance() access-violates. Removed; the OS
  reclaims the library at process exit either way.
- Exit-time heap corruption (0xC0000409) once a backend actually holds
  live device/queue state (OneAPI, once it can load at all -- see above):
  tenzor::finalize() was only ever invoked via the C++ library's own
  std::atexit(), which races against other DLLs' (the SYCL runtime's) own
  exit-time cleanup in the shared process-wide atexit cascade. Exposed
  finalize() to Python and register it via Python's own atexit module, so
  it runs deterministically while the interpreter is still alive.
  Verified with a minimal non-Python repro: explicit finalize() before
  main() returns never crashes; relying only on the atexit-registered
  call always did once OneAPI had real devices.
- cudaGetErrorString()/hipGetErrorString() can return null on a machine
  with no vendor driver installed at all (not just no GPU) -- CUDA's
  device_count() passed the result straight into an fmt {} placeholder
  (threw "string pointer is null", caught only by spdlog's generic error
  handler); ROCm's constructor concatenated it onto a std::string
  (undefined behavior on null, not even a catchable exception). Both
  guarded with a fallback string.

Docs
- CLAUDE.md: add a Windows section covering the Ninja-not-VS-generator
  requirement and how DLL loading is (and isn't) handled automatically.
- INSTALL.md: replace the untested "Visual Studio generator" Windows
  instructions with the actual working Ninja + MSVC + vcpkg process.
- .gitignore: /build_*/ alongside the existing /build-*/ pattern.

Also fixes multi-generation UTF-8/CP1252 mojibake corruption (introduced
by tooling in an earlier session, not content changes) in comments across
19 files -- em dashes and math symbols in kernel comments had been
silently re-encoded through 1-3 rounds of UTF-8-as-CP1252 misinterpretation.

+++ src/backends/cpu/kernels/simd_elementwise.hpp
@@ -64,11 +64,11 @@ void elementwise_unary(const T* input, T* output, size_t n, Op op) {
     if constexpr (std::is_same_v<T, float> || std::is_same_v<T, double>) {
         if (static_cast<int64_t>(n) >= ELEMENTWISE_OMP_THRESHOLD) {
             #pragma omp parallel for schedule(static)
-            for (size_t i = 0; i < n; ++i) {
+            for (int64_t i = 0; i < n; ++i) {
                 output[i] = op(input[i]);
             }
         } else {
-            for (size_t i = 0; i < n; ++i) {
+            for (int64_t i = 0; i < n; ++i) {
                 output[i] = op(input[i]);
             }
         }
@@ -101,15 +101,15 @@ void elementwise_unary(const T* input, T* output, size_t n, Op op) {
         };
         if (static_cast<int64_t>(n) >= ELEMENTWISE_OMP_THRESHOLD) {
             #pragma omp parallel for schedule(static)
-            for (size_t chunk = 0; chunk < n; chunk += 512) {
-                process_chunk(chunk, std::min(chunk + 512, n));
+            for (int64_t chunk = 0; chunk < n; chunk += 512) {
+                process_chunk(chunk, std::min(chunk + 512, static_cast<int64_t>(n)));
             }
         } else {
             process_chunk(0, n);
         }
 #else
         // Scalar fallback without F16C
-        for (size_t i = 0; i < n; ++i) {
+        for (int64_t i = 0; i < n; ++i) {
             float val = static_cast<float>(input[i]);
             output[i] = static_cast<T>(op(val));
         }
@@ -144,21 +144,21 @@ void elementwise_unary(const T* input, T* output, size_t n, Op op) {
         };
         if (static_cast<int64_t>(n) >= ELEMENTWISE_OMP_THRESHOLD) {
             #pragma omp parallel for schedule(static)
-            for (size_t chunk = 0; chunk < n; chunk += 512) {
-                process_chunk(chunk, std::min(chunk + 512, n));
+            for (int64_t chunk = 0; chunk < n; chunk += 512) {
+                process_chunk(chunk, std::min(chunk + 512, static_cast<int64_t>(n)));
             }
         } else {
             process_chunk(0, n);
         }
 #else
-        for (size_t i = 0; i < n; ++i) {
+        for (int64_t i = 0; i < n; ++i) {
             float val = static_cast<float>(input[i]);
             output[i] = static_cast<T>(op(val));
         }
 #endif
     } else {
         // Other types: scalar fallback
-        for (size_t i = 0; i < n; ++i) {
+        for (int64_t i = 0; i < n; ++i) {
             float val = static_cast<float>(input[i]);
             output[i] = static_cast<T>(op(val));
         }
@@ -178,11 +178,11 @@ void elementwise_binary(const T* a, const T* b, T* output, size_t n, Op op) {
     if constexpr (std::is_same_v<T, float> || std::is_same_v<T, double>) {
         if (static_cast<int64_t>(n) >= ELEMENTWISE_OMP_THRESHOLD) {
             #pragma omp parallel for schedule(static)
-            for (size_t i = 0; i < n; ++i) {
+            for (int64_t i = 0; i < n; ++i) {
                 output[i] = op(a[i], b[i]);
             }
         } else {
-            for (size_t i = 0; i < n; ++i) {
+            for (int64_t i = 0; i < n; ++i) {
                 output[i] = op(a[i], b[i]);
             }
         }
@@ -219,14 +219,14 @@ void elementwise_binary(const T* a, const T* b, T* output, size_t n, Op op) {
         };
         if (static_cast<int64_t>(n) >= ELEMENTWISE_OMP_THRESHOLD) {
             #pragma omp parallel for schedule(static)
-            for (size_t chunk = 0; chunk < n; chunk += 512) {
-                process_chunk(chunk, std::min(chunk + 512, n));
+            for (int64_t chunk = 0; chunk < n; chunk += 512) {
+                process_chunk(chunk, std::min(chunk + 512, static_cast<int64_t>(n)));
             }
         } else {
             process_chunk(0, n);
         }
 #else
-        for (size_t i = 0; i < n; ++i) {
+        for (int64_t i = 0; i < n; ++i) {
             float va = static_cast<float>(a[i]);
             float vb = static_cast<float>(b[i]);
             output[i] = static_cast<T>(op(va, vb));
@@ -265,21 +265,21 @@ void elementwise_binary(const T* a, const T* b, T* output, size_t n, Op op) {
         };
         if (static_cast<int64_t>(n) >= ELEMENTWISE_OMP_THRESHOLD) {
             #pragma omp parallel for schedule(static)
-            for (size_t chunk = 0; chunk < n; chunk += 512) {
-                process_chunk(chunk, std::min(chunk + 512, n));
+            for (int64_t chunk = 0; chunk < n; chunk += 512) {
+                process_chunk(chunk, std::min(chunk + 512, static_cast<int64_t>(n)));
             }
         } else {
             process_chunk(0, n);
         }
 #else
-        for (size_t i = 0; i < n; ++i) {
+        for (int64_t i = 0; i < n; ++i) {
             float va = static_cast<float>(a[i]);
             float vb = static_cast<float>(b[i]);
             output[i] = static_cast<T>(op(va, vb));
         }
 #endif
     } else {
-        for (size_t i = 0; i < n; ++i) {
+        for (int64_t i = 0; i < n; ++i) {
             float va = static_cast<float>(a[i]);
             float vb = static_cast<float>(b[i]);
             output[i] = static_cast<T>(op(va, vb));
